requirements.txt
Django
djangorestframework
python-decouple==3.5
requests
gunicorn
django-heroku